refactor: agent workspace dependencies
