+++ agent/Cargo.toml
@@ -8,7 +8,13 @@ default-run = "deepflow-agent"
 
 [workspace]
 members = ["crates/*", "plugins/packet_segmentation_reassembly"]
-exclude = ["src/ebpf/samples/rust/profiler", "src/ebpf/samples/rust/socket-tracer"]
+exclude = [
+    "src/ebpf/samples/rust/profiler",
+    "src/ebpf/samples/rust/socket-tracer",
+]
+
+[workspace.dependencies]
+prost = "0.11"
 
 [dependencies]
 ahash = "0.8"
@@ -66,7 +72,7 @@ parking_lot = "0.11"
 pcap_assembler = { path = "plugins/pcap_assembler" }
 pcap-sys = "0.1.3"
 pnet = "^0.29"
-prost = "0.11"
+prost.workspace = true
 public = { path = "crates/public" }
 rand = "0.8.5"
 regex = "1"
