+++ agent/crates/public/Cargo.toml
@@ -18,7 +18,7 @@ log = "0.4"
 num_enum = "0.5.6"
 parking_lot = "0.11"
 pnet = "^0.29"
-prost = "0.11"
+prost.workspace = true
 regex = "1"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0.72"
