+++ agent/plugins/packet_sequence_block/Cargo.toml
@@ -6,5 +6,5 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-prost = "0.11.0"
-public = { path = "../../crates/public"}
+prost.workspace = true
+public = { path = "../../crates/public" }
